Fix remaining TDZ circular imports, re-enable journey timeline
Two more circular dependency cycles were causing 'Cannot access before
initialization' at runtime:

Cycle A: relationshipGraph → plannerGroceryUtils → relationshipGraph
  Fix: move normalizeMealIngredient + its private helpers into a new
  ingredientNormalization.ts with no upward deps. Both files now import
  from there; the mutual dependency is gone.

Cycle B: plannerObjectiveEngine → plannerObjectiveScoring
         → autoPlannerTradeoffAnalysis → plannerObjectiveEngine
  Fix: remove the evaluatePlannerObjectives import from
  autoPlannerTradeoffAnalysis (calculateCompositeOptimizationScore now
  uses the legacy score exclusively). Also switch the metrics imports in
  that file to autoPlannerMetrics to avoid the re-export chain.

ENABLE_JOURNEY_TIMELINE = true — timeline is fully enabled.
