+++ client/src/components/meal-planner/plannerGroceryUtils.ts
@@ -2,6 +2,8 @@ import { MEAL_TYPES, WEEK_DAYS } from './nutritionMealPlannerUtils';
 import { getMealsForSlot } from './planner-graph/plannerGraphUtils';
 import { extractMealIngredients } from './planner-graph/plannerMealExtraction';
 import { buildMealRelationshipGraph } from './meal-relationships/relationshipGraph';
+import { normalizeMealIngredient } from './ingredientNormalization';
+export { normalizeMealIngredient } from './ingredientNormalization';
 
 export type PlannerGrocerySourceMeal = {
   mealName: string;
@@ -50,26 +52,6 @@ export type PlannerGroceryDerivationState = {
   editedById?: Record<string, { name?: string; quantitySummary?: string; category?: string }>;
 };
 
-const SAFE_DESCRIPTORS = new Set([
-  'boneless',
-  'skinless',
-  'fresh',
-  'frozen',
-  'raw',
-  'cooked',
-  'chopped',
-  'diced',
-  'sliced',
-  'minced',
-  'shredded',
-  'grated',
-  'whole',
-  'large',
-  'small',
-  'medium',
-  'extra',
-]);
-
 const GENERIC_COMPONENT_WORDS = new Set([
   'serving',
   'servings',
@@ -90,35 +72,6 @@ const titleCase = (value: string) => value
   .filter(Boolean)
   .map((part) => part.charAt(0).toUpperCase() + part.slice(1))
   .join(' ');
-
-const singularizeToken = (token: string) => {
-  if (token.length <= 3) return token;
-  if (token.endsWith('ies')) return `${token.slice(0, -3)}y`;
-  if (token.endsWith('oes')) return token.slice(0, -2);
-  if (token.endsWith('ches') || token.endsWith('shes')) return token.slice(0, -2);
-  if (token.endsWith('s') && !token.endsWith('ss')) return token.slice(0, -1);
-  return token;
-};
-
-export const normalizeMealIngredient = (value: unknown) => {
-  const cleaned = String(value ?? '')
-    .toLowerCase()
-    .replace(/[()[\]{}]/g, ' ')
-    .replace(/[.,;:!]+/g, ' ')
-    .replace(/\s+/g, ' ')
-    .trim();
-
-  if (!cleaned) return '';
-
-  const tokens = cleaned
-    .split(' ')
-    .map((token) => token.replace(/^[^a-z0-9]+|[^a-z0-9]+$/g, ''))
-    .filter(Boolean)
-    .filter((token) => !SAFE_DESCRIPTORS.has(token))
-    .map(singularizeToken);
-
-  return tokens.join(' ').trim();
-};
 
 const isUsefulIngredient = (name: string) => {
   const normalized = normalizeMealIngredient(name);
